:sparkles: Add monitor lock timeout function

+++ monitor_ctdb.sh
@@ -23,10 +23,20 @@ function get_all_banned_count() {
 }
 
 function get_lock_del_interval() {
-    last_time_info=$(sed -n 2p $STATUS_FILE)
-    last_time=${last_time_info#*:}
+    last_time=$(get_lock_del_last_time)
     current_time=$(date +%s)
     interval=$(expr $current_time - $last_time)
+    echo $interval
+}
+
+function get_ctdb_restart_interval() {
+    last_time=$(get_ctdb_restart_last_time)
+    if [ -z "$last_time" ]; then
+        interval=$(expr $RESTART_CTDB_INTERVAL_MAX + 1)
+    else
+        current_time=$(date +%s)
+        interval=$(expr $current_time - $last_time)
+    fi
     echo $interval
 }
 
@@ -36,17 +46,23 @@ function get_lock_del_last_time() {
     echo $last_time
 }
 
+function get_ctdb_restart_last_time() {
+    last_time_info=$(sed -n 3p $STATUS_FILE)
+    last_time=${last_time_info#*:}
+    echo $last_time
+}
+
 function update_count() {
     update_type=$1
     if [ $update_type = "new" ];then
         count=$(get_all_banned_count)
-        new_count=$(expr $count + 1)
+        new_count=$(expr $count + 1) 
     elif [ $update_type = "reset" ];then
         new_count=0
     fi
     count_info="all banned count:"$new_count
     sed -i 1s/.*/"$count_info"/ $STATUS_FILE
-    echo $(date)" Ctdb all banned count: "$count >> $MONITOR_LOG
+    echo $(date)" Ctdb all banned count: "$new_count >> $MONITOR_LOG
 }
 
 function update_last_del_time() {
@@ -54,12 +70,19 @@ function update_last_del_time() {
     current_time=$(date +%s)
     if [ $update_type = "new" ];then
         time_info="lock last delete time:"$current_time
-    	echo $(date)" Ctdb last delete time update: "$current_time >> $MONITOR_LOG
+    	echo $(date)" Lock last delete time update: "$current_time >> $MONITOR_LOG
     elif [ $update_type = "reset" ];then
         time_info="lock last delete time:"
-    	echo $(date)" Ctdb last delete time reset" >> $MONITOR_LOG
+    	echo $(date)" Lock last delete time reset" >> $MONITOR_LOG
     fi
     sed -i 2s/.*/"$time_info"/ $STATUS_FILE
+}
+
+function update_last_ctdb_restart_time() {
+    current_time=$(date +%s)
+    time_info="ctdb last restart time:"$current_time
+    echo $(date)" Ctdb last restart time update: "$current_time >> $MONITOR_LOG
+    sed -i 3s/.*/"$time_info"/ $STATUS_FILE
 }
 
 function delete_lock_object() {
@@ -74,6 +97,7 @@ function delete_lock_object() {
 function generate_status_file() {
     echo "all banned count:0" > $STATUS_FILE
     echo "lock last delete time:" >> $STATUS_FILE
+    echo "ctdb last restart time:" >> $STATUS_FILE
 }
 
 function save_nodes_ip() {
@@ -86,12 +110,21 @@ function save_nodes_ip() {
     done
 }
 
-function monitor_master() {
-    ctdb_status=$(ctdb status 2>&1)
-
+function check_status_file() {
     if [ ! -f "$STATUS_FILE" ]; then
         generate_status_file
     fi
+    line1=$(sed -n 1p $STATUS_FILE)
+    line2=$(sed -n 2p $STATUS_FILE)
+    line3=$(sed -n 3p $STATUS_FILE)
+    if [ -z "$line1" -o -z "$line2" -o -z "$line3" ]; then
+        generate_status_file
+    fi
+}
+
+function monitor_master() {
+    check_status_file
+    ctdb_status=$(ctdb status 2>&1)
 
     if [ "$ctdb_status" = "$ALL_BANNED" ]; then
         update_count new
@@ -100,13 +133,15 @@ function monitor_master() {
             if [ -z $(get_lock_del_last_time) ]; then
                 delete_lock_object
             else
-                if [ $(get_lock_del_interval) -ge $INTERVAL_MAX ]; then
+                if [ $(get_lock_del_interval) -ge $DEL_LOCK_INTERVAL_MAX ]; then
                     delete_lock_object
                 fi
             fi
         fi
     else
-        update_count reset
+        if [ $(get_all_banned_count) -ne 0 ]; then
+            update_count reset
+        fi
     fi
 }
 
@@ -115,10 +150,13 @@ function monitor_nodes_ip_in_rados() {
     for ipinfo in $ips; do
         ip=${ipinfo%/*}
         if $(timeout 10 rados -p rbd ls | grep "$ip" -qw); then
-            systemctl restart ctdb
-            echo $(date)" Ctdb restart" >> $MONITOR_LOG
-            rados -p rbd rm $ip
-            echo $(date)" Delete ip in rados: "$ip >> $MONITOR_LOG
+            if [ $(get_ctdb_restart_interval) -ge $RESTART_CTDB_INTERVAL_MAX ]; then 
+                systemctl restart ctdb
+                echo $(date)" Ctdb restart by delete ip in rados" >> $MONITOR_LOG
+                rados -p rbd rm $ip
+                echo $(date)" Delete ip in rados: "$ip >> $MONITOR_LOG
+                update_last_ctdb_restart_time
+            fi
         fi
     done
 }
@@ -129,7 +167,15 @@ function get_current_node_ips() {
 }
 
 function monitor_get_lock_timeout() {
-    rados_helper_count=$()
+    count=$(ps -ef | grep rados_helper | wc -l)
+    if [ $count -ge $RADOS_HELPER_PROCESS_MAX ]; then
+        echo $(date)" Rados_helper process redundant, count is "$count >> $MONITOR_LOG
+        if [ $(get_ctdb_restart_interval) -ge $RESTART_CTDB_INTERVAL_MAX ]; then 
+            systemctl restart ctdb
+            echo $(date)" Ctdb restart by redundant rados_helper process" >> $MONITOR_LOG
+            update_last_ctdb_restart_time
+        fi
+    fi
 }
 
 CTDB_CONFIG_FILE=/etc/sysconfig/ctdb
@@ -137,11 +183,14 @@ STATUS_FILE=/etc/ctdb/status.log
 MONITOR_LOG=/var/log/monitor_ctdb.log
 ALL_BANNED="Warning: All nodes are banned."
 COUNT_MAX=5
-INTERVAL_MAX=300
+DEL_LOCK_INTERVAL_MAX=300
+RADOS_HELPER_PROCESS_MAX=6
+RESTART_CTDB_INTERVAL_MAX=120
 
 if $(grep rados $CTDB_CONFIG_FILE -q); then
     if $(check_if_master); then
 	monitor_master
     fi
     monitor_nodes_ip_in_rados
+    monitor_get_lock_timeout
 fi
